Проектирование и разработка информационных систем/src/Пример для PY/image_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pillow==11.3.0
python-multipart==0.0.6